+++ src/components/shared/flight/FlightList/FlightList.test.tsx
@@ -1,31 +1,16 @@
 import { FlightList } from '@/components/shared/flight/FlightList/FlightList';
-import { createClient } from '@/datx/create-client';
-import { buildMany, createFactory, sequence } from '@datx/test-data-factory';
-import { Flight } from '@/models/Flight';
+import { buildMany } from '@datx/test-data-factory';
+
 import { render, screen } from 'test-utils';
-
-const client = createClient();
-const factory = createFactory(client);
-
-const flight = factory(Flight, {
-	fields: {
-		id: sequence(),
-		name: 'Air Force One',
-		arrivesAt: '2021-01-01T00:00:00.000Z',
-		airplaneModel: 'Boeing 747',
-		basePrice: '100',
-		currentSeatPrice: '100',
-		departsAt: '2021-01-01T00:00:00.000Z',
-	},
-});
+import { flightFactory } from '__mocks__/factories';
 
 describe('FlightList', () => {
 	beforeEach(() => {
-		factory.reset();
+		flightFactory.reset();
 	});
 
 	it('should render flights', () => {
-		const data = buildMany(flight, 2);
+		const data = buildMany(flightFactory, 2);
 
 		render(<FlightList flightList={data} />);
